--- src/recoil/atoms/retryCountState.ts
@@ -1,6 +1,0 @@
-import { atom } from 'recoil';
-
-export const retryCountState = atom({
-  key: 'retryCountState',
-  default: 0,
-});
